MNT: move simulated devices to separate submodule
